requirements.txt
fastapi
uvicorn
pydantic
aiohttp
pandas
python-multipart
python-jose[cryptography]
passlib[bcrypt]
openai
httpx
asyncio
sqlalchemy
alembic
python-dotenv
pydantic-settings
anthropic
aiofiles
aiosqlite
sentence-transformers
numpy
faiss-cpu

